# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sales Support API",
        "description": "SES企業向け営業支援システム API",
        "contact": {
            "email": "admin@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "ローカル開発環境"
        }
    ],
    "paths": {
...
        "/api/v1/emails": {
            "get": {
                "tags": [
                    "Emails"
                ],
                "summary": "メール一覧取得",
                "operationId": "6c3485ecb0f17fdfbe796aa0ab2934d6",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "件名・送信者・本文で検索",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "unread",
                        "in": "query",
                        "description": "未読のみ",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "category",
                        "in": "query",
                        "description": "カテゴリ",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "engineer",
                                "project"
                            ]
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "1ページ件数",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 30
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功"
                    },
                    "401": {
                        "description": "認証エラー"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/emails/{id}": {
            "get": {
                "tags": [
                    "Emails"
                ],
                "summary": "メール詳細取得（自動既読）",
                "operationId": "03ab37d74a110a1d3f447afc4141e49c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "メールID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功"
                    },
                    "404": {
                        "description": "見つかりません"
                    },
                    "401": {
                        "description": "認証エラー"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/emails/sync": {
            "post": {
                "tags": [
                    "Emails"
                ],
                "summary": "Gmailから最新メールを同期",
                "operationId": "aed3ee3a1a946972c3aa06693693032d",
                "responses": {
                    "200": {
                        "description": "同期成功"
                    },
                    "401": {
                        "description": "Gmailトークン失効"
                    },
                    "422": {
                        "description": "Gmail未接続"
                    },
                    "500": {
                        "description": "同期失敗"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/emails/{id}/link": {
            "patch": {
                "tags": [
                    "Emails"
                ],
                "summary": "メールを担当者・商談・顧客に紐付け",
                "operationId": "ead77b705f6f6579f246b6304ad5ce26",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "メールID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功"
                    },
                    "401": {
                        "description": "認証エラー"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/emails/unread-count": {
            "get": {
                "tags": [
                    "Emails"
                ],
                "summary": "未読メール件数取得",
                "operationId": "4f5d96d975403d70199bb1b53eda428f",
                "responses": {
                    "200": {
                        "description": "成功"
                    },
                    "401": {
                        "description": "認証エラー"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/emails/mark-all-read": {
            "post": {
                "tags": [
                    "Emails"
                ],
                "summary": "全メールを既読にする",
                "operationId": "951808147ba27510d03081c86ffedaa4",
                "responses": {
                    "200": {
                        "description": "成功"
                    },
                    "401": {
                        "description": "認証エラー"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Supabase Auth JWT トークン",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}